requirements.txt
# Core dependencies
aiohttp>=3.8.0
pydantic>=1.10.0
pyyaml>=6.0
rdflib>=6.0.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.990

